feat: begin adding accounts
While having a publicly accessible website is great and all, it means
anyone can view my items and modify them. It's also not very extensible
if others feel like giving it a try.

Let's start working in the concept of accounts, so that we can create
them manually in the database and then associate all the existing items
with the first one.

This change:
* Adds a basic table structure
* Adds some persistence and tests around it